I want to make the table bigger. The days should be at the top and the rest on the left side. I would also like to create a popup fill method to make it easier to enter data in the table. Please add material design. The operation of the printing of the pdf, logo, titles, and footer should not be touched. I will use the supplied code for the improvements. Everything is based on HTML.
Here are the specific changes that I would make to the table:

I would increase the width of the table to make it easier to read.
I would move the days to the top of the table and the rest of the data to the left side. This would make it easier to scan the table and find the information that I am looking for.
I would create a popup fill method that would allow me to enter data in the table without having to scroll up and down. This would make it easier to enter large amounts of data.
I would add material design to the table to make it look more modern and professional.
I would not touch the operation of the printing of the pdf, logo, titles, or footer. These elements are important for the overall look and feel of the table.