Here's a summary of the changes made:
Added a title to the cell containing the "Church Services" section with the id "services-title".
Added a welcome message to the page.
Reduced the size of the "Church Services" cell by adding padding to the title.
Fixed the issue with the pop-up form not saving information by checking if there is a focused cell before saving the data.
Made minor corrections and improvements based on the existing code.
Please note that this code assumes you have the necessary CSS and JavaScript dependencies included for Materialize CSS framework.
If you encounter any issues, make sure to include those dependencies correctly.