Added a welcome message: A paragraph <p> element with the welcome message was added below the subheadings.
Reduced the width of the cells in the first column: The CSS class
.first-column was removed, and a new CSS rule was added to reduce
the width of the cells in the first column to 20%.

Fixed the issue with the pop-up form not saving information:
The save functionality was implemented by adding an event listener
to the modal's save button. When the button is clicked,
the value in the modal's input field is saved to the active cell
that was being edited.